sidecars/QuantumSidecar/requirements.txt
qiskit>=1.0.0
qiskit-aer>=0.14.0
qiskit-ibm-runtime>=0.20.0
amazon-braket-sdk>=1.70.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
numpy>=1.26.0
